fix(coordination): address PR #416 review feedback
- DatabaseMembershipStoreBase: ReadSnapshot -> async ReadSnapshotAsync via DbDataReader async column APIs
- Coordination serializer: inject IJsonSerializer, registered as a keyed service (CoordinationOptions.JsonSerializerServiceKey) so consumers can override coordination serialization independently
- CoordinationOptions: source-generated [GeneratedRegex] for cluster-name validation
- MembershipService: use CancellationTokenSource.CancelAsync() in async _SignalLocalMembershipLostAsync
- Redis scripts: one file per RedisScriptDefinition with co-located params struct; relocate single-consumer scripts to owning packages (DistributedLocks.Redis, Caching.Redis, Coordination.Redis); RemoveIfEqual/ReplaceIfEqual stay shared in Headless.Redis; suppress IDE1006 on Lua @param structs

Build clean (0 warn/err); 257 integration tests pass across Redis/DistributedLocks.Redis/Caching.Redis/Coordination.{Redis,PostgreSql,SqlServer}.
