+++ src/Headless.Coordination.PostgreSql/PostgresMembershipStore.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Mahmoud Shaheen. All rights reserved.
 
 using System.Data;
+using Headless.Serializer;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Npgsql;
 using NpgsqlTypes;
@@ -10,8 +12,9 @@ namespace Headless.Coordination.PostgreSql;
 #pragma warning disable CA2100 // SQL text is built from internal schema constants only.
 internal sealed class PostgresMembershipStore(
     IOptions<PostgreSqlCoordinationOptions> providerOptions,
-    IOptions<CoordinationOptions> coordinationOptions
-) : DatabaseMembershipStoreBase(coordinationOptions.Value)
+    IOptions<CoordinationOptions> coordinationOptions,
+    [FromKeyedServices(CoordinationOptions.JsonSerializerServiceKey)] IJsonSerializer serializer
+) : DatabaseMembershipStoreBase(coordinationOptions.Value, serializer)
 {
     protected override async ValueTask<NodeIncarnation> AllocateIncarnationCoreAsync(
         string clusterName,
@@ -251,7 +254,7 @@ internal sealed class PostgresMembershipStore(
 
         while (await reader.ReadAsync(cancellationToken).ConfigureAwait(false))
         {
-            snapshots.Add(ReadSnapshot(reader));
+            snapshots.Add(await ReadSnapshotAsync(reader, cancellationToken).ConfigureAwait(false));
         }
 
         return snapshots;
