+++ src/Headless.Coordination.Redis/Scripts/RedisMembershipHeartbeatScriptDefinition.cs
@@ -1,0 +1,52 @@
+// Copyright (c) Mahmoud Shaheen. All rights reserved.
+
+using System.Runtime.InteropServices;
+using Headless.Redis;
+using StackExchange.Redis;
+
+namespace Headless.Coordination.Redis;
+
+/// <summary>Stores a guarded coordination heartbeat using Redis server time.</summary>
+internal sealed class RedisMembershipHeartbeatScriptDefinition : RedisScriptDefinition
+{
+    public static RedisMembershipHeartbeatScriptDefinition Instance { get; } = new();
+
+    private RedisMembershipHeartbeatScriptDefinition()
+        : base(
+            """
+            local current = redis.call('get', @genKey)
+            if current == false or tonumber(current) ~= tonumber(@incarnation) then
+              return 0
+            end
+
+            local nowSecMicro = redis.call('TIME')
+            local nowMs = (tonumber(nowSecMicro[1]) * 1000) + math.floor(tonumber(nowSecMicro[2]) / 1000)
+            local hardExpiryMs = nowMs + tonumber(@hardMs)
+            local payload = cjson.encode({
+              last_beat_ms = nowMs,
+              role = @role,
+              metadata = @metadata
+            })
+
+            redis.call('zadd', @liveKey, hardExpiryMs, @member)
+            redis.call('hset', @knownKey, @member, payload)
+
+            return 1
+            """
+        ) { }
+}
+
+#pragma warning disable IDE1006 // camelCase mirrors the Lua @param token names
+/// <summary>Parameters for <see cref="RedisMembershipHeartbeatScriptDefinition"/>.</summary>
+[StructLayout(LayoutKind.Auto)]
+internal readonly record struct HeartbeatParams(
+    RedisKey liveKey,
+    RedisKey knownKey,
+    RedisKey genKey,
+    string member,
+    long incarnation,
+    long hardMs,
+    string role,
+    string metadata
+);
+#pragma warning restore IDE1006
