+++ src/Headless.Redis/README.md
@@ -14,7 +14,7 @@ Provides Redis helper extensions plus definition-first Lua script loading/execut
 
 ## Design Notes
 
-`Headless.Redis` owns script definitions and generic loading only. Provider packages own script grouping, hosted warmup, typed parameters, and result decoding so consumers load only the script definitions they need.
+`Headless.Redis` owns the generic loader plus a small set of broadly shared CAS script definitions (`RemoveIfEqualScriptDefinition`, `ReplaceIfEqualScriptDefinition`). Each provider package owns its own script definitions, script grouping, hosted warmup, typed parameters, and result decoding so consumers load only the script definitions they need. Scripts used by a single provider live in that provider's package (for example, the cache CAS and counter scripts live in `Headless.Caching.Redis`, the lock/semaphore scripts in `Headless.DistributedLocks.Redis`, and the membership scripts in `Headless.Coordination.Redis`).
 
 Each concrete `RedisScriptDefinition` type is a singleton contract. Reuse the exposed `Instance` member; the loader rejects multiple instances of the same concrete type because it caches loaded scripts by definition type.
 
@@ -32,7 +32,7 @@ var builder = WebApplication.CreateBuilder(args);
 var redis = await ConnectionMultiplexer.ConnectAsync("localhost");
 var scriptsLoader = new HeadlessRedisScriptsLoader(redis);
 
-await scriptsLoader.LoadAsync([IncrementWithExpireScriptDefinition.Instance]);
+await scriptsLoader.LoadAsync([ReplaceIfEqualScriptDefinition.Instance]);
 ```
 
 ## Usage
@@ -43,13 +43,13 @@ await scriptsLoader.LoadAsync([IncrementWithExpireScriptDefinition.Instance]);
 var db = redis.GetDatabase();
 var result = await scriptsLoader.EvaluateAsync(
     db,
-    IncrementWithExpireScriptDefinition.Instance,
-    new
-    {
-        key = (RedisKey)"counter",
-        value = (RedisValue)1,
-        expires = 60_000,
-    }
+    ReplaceIfEqualScriptDefinition.Instance,
+    new ReplaceIfEqualParams(
+        key: (RedisKey)"counter",
+        value: "2",
+        expected: "1",
+        expires: 60_000
+    )
 );
 ```
