+++ src/Headless.Redis/Scripts/ReplaceIfEqualScriptDefinition.cs
@@ -1,0 +1,36 @@
+// Copyright (c) Mahmoud Shaheen. All rights reserved.
+
+using System.Runtime.InteropServices;
+using StackExchange.Redis;
+
+namespace Headless.Redis;
+
+/// <summary>Atomically replaces a value only if it matches the expected value.</summary>
+public sealed class ReplaceIfEqualScriptDefinition : RedisScriptDefinition
+{
+    public static ReplaceIfEqualScriptDefinition Instance { get; } = new();
+
+    private ReplaceIfEqualScriptDefinition()
+        : base(
+            """
+            local currentVal = redis.call('get', @key)
+            local expected = @expected
+            if expected == '' then expected = false end
+            if currentVal == expected then
+              if (@expires ~= nil and @expires ~= '') then
+                return redis.call('set', @key, @value, 'PX', @expires) and 1 or 0
+              else
+                return redis.call('set', @key, @value) and 1 or 0
+              end
+            else
+              return -1
+            end
+            """
+        ) { }
+}
+
+#pragma warning disable IDE1006 // camelCase mirrors the Lua @param token names
+/// <summary>Parameters for <see cref="ReplaceIfEqualScriptDefinition"/>.</summary>
+[StructLayout(LayoutKind.Auto)]
+public readonly record struct ReplaceIfEqualParams(RedisKey key, string? value, string expected, RedisValue expires);
+#pragma warning restore IDE1006
